Educational Application using Machine and Deep Learning/requirements.txt
appdirs==1.4.1
click==6.7
Flask==1.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==0.23
numpy==1.12.0
packaging==16.8
protobuf==3.2.0
pyparsing==2.1.10
six==1.10.0
tensorflow
Werkzeug
scipy
Pillow
keras
h5py
